refactor: extract ImageCheckboxes
