+++ pages/quiz/how-are-you-feeling.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react'
 import styled from 'styled-components'
-import { NextRouter, useRouter } from 'next/router'
+import { useRouter } from 'next/router'
 
 import { Layout } from '../../src/Components/Layout'
 import { HeaderComponent } from '../../src/Components/Header'
@@ -8,160 +8,7 @@ import { VerticalSpacing } from '../../src/Components/VerticalSpacing'
 import { QuizContext } from '../../src/context/QuizContext'
 import { LinkButton } from '../../src/Components/LinkButton'
 import { StickyNavBar } from '../../src/Components/StickyNavBar'
-
-interface ImageCheckboxesProps {
-  id: string
-  values: readonly { title: Feeling; image: string }[]
-  label: string
-  router: NextRouter
-}
-
-// Wraps the checkbox area of component
-const CheckboxGroup = styled.div`
-  legend {
-    border-bottom: 1px solid ${(props) => props.theme.colours.yellow};
-    font-family: 'Catamaran', sans-serif;
-    font-size: ${(props) => props.theme.fontSizes.heading};
-  }
-
-  fieldset {
-    border: none;
-    padding: 0;
-  }
-`
-
-// Contains all the CheckboxItems
-const CheckboxContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  margin: 0 -0.5rem;
-`
-
-// This is the visual representation of the checked status
-const CheckedDisplay = styled.div`
-  border-radius: 1rem;
-  border-top-right-radius: 2rem;
-  height: 100%;
-  position: absolute;
-  width: 100%;
-`
-
-// This is the text of the checkbox item
-const Text = styled.div`
-  align-items: center;
-  display: flex;
-  font-size: ${(props) => props.theme.fontSizes.extraSmall};
-  font-weight: bold;
-  justify-content: center;
-  letter-spacing: 0.05em;
-  margin: 0 0.5rem 0.5rem 0.5rem;
-  z-index: 1;
-`
-
-// Contains label, input, image
-const CheckboxItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  flex: 0 0 33%;
-  padding: 0 0.5rem;
-  margin: 0.5rem 0;
-
-  label {
-    border-radius: 1rem;
-    border-top-right-radius: 2rem;
-    box-shadow: 0 0 9px 2px ${(props) => props.theme.colours.shadow};
-    cursor: pointer;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    text-align: center;
-  }
-
-  input {
-    appearance: none;
-    background: none;
-    height: 1px;
-    opacity: 0.00001;
-    position: absolute;
-    width: 100%;
-    cursor: pointer;
-
-    &:focus ~ ${CheckedDisplay} {
-      outline: 3px dashed ${(props) => props.theme.colours.blue};
-      outline-offset: 5px;
-    }
-
-    &:checked ~ ${CheckedDisplay} {
-      background-color: ${(props) => props.theme.colours.aqua_light};
-    }
-
-    &:checked ~ ${Text} {
-      color: ${(props) => props.theme.colours.purple};
-    }
-
-    &:checked ~ img {
-      border-radius: 50%;
-    }
-  }
-
-  img {
-    border-radius: 1rem;
-    border: 5px solid transparent;
-    height: auto;
-    margin: auto;
-    margin-top: 5px;
-    width: calc(100% - 10px);
-    z-index: 1;
-  }
-`
-
-export const ImageCheckboxes = ({
-  id,
-  values,
-  label,
-  router,
-}: ImageCheckboxesProps): JSX.Element => {
-  const { howAreFeelingGet, howAreFeelingToggle } = useContext(QuizContext)
-
-  return (
-    <>
-      <CheckboxGroup>
-        <fieldset>
-          <legend>{label}</legend>
-          <VerticalSpacing size={1} />
-          <LinkButton
-            textContent="skip this question"
-            arrow="right"
-            onClick={() => router.push('what-are-your-interests')}
-          />
-          <VerticalSpacing />
-          <CheckboxContainer>
-            {values.map((checkbox) => {
-              return (
-                <CheckboxItem key={checkbox.title}>
-                  <label>
-                    <input
-                      type="checkbox"
-                      name={id}
-                      value={checkbox.title}
-                      checked={howAreFeelingGet(checkbox.title)}
-                      onChange={() => howAreFeelingToggle(checkbox.title)}
-                    />
-                    <img src={checkbox.image} alt="" />
-                    <Text>{checkbox.title}</Text>
-                    <CheckedDisplay />
-                  </label>
-                </CheckboxItem>
-              )
-            })}
-          </CheckboxContainer>
-        </fieldset>
-      </CheckboxGroup>
-      <VerticalSpacing />
-    </>
-  )
-}
+import ImageCheckboxes from '../../src/Components/ImageCheckboxes'
 
 const Section = styled.section`
   align-items: center;
@@ -215,6 +62,20 @@ const StyledLink = styled.a`
   }
 `
 
+// Wraps the checkbox area of component
+const CheckboxGroup = styled.div`
+  legend {
+    border-bottom: 1px solid ${(props) => props.theme.colours.yellow};
+    font-family: 'Catamaran', sans-serif;
+    font-size: ${(props) => props.theme.fontSizes.heading};
+  }
+
+  fieldset {
+    border: none;
+    padding: 0;
+  }
+`
+
 const feelings = [
   { title: 'unsure', image: '/img/unsure.svg' },
   { title: 'ok', image: '/img/ok.svg' },
@@ -234,7 +95,7 @@ export type Feeling = typeof feelings[number]['title']
 
 export const HowAreYouFeelingPage = (): JSX.Element => {
   const router = useRouter()
-  const { howAreFeelingGet } = useContext(QuizContext)
+  const { howAreFeelingGet, howAreFeelingToggle } = useContext(QuizContext)
 
   const nextPage = () => {
     const triggerFeelings: Array<Feeling> = [
@@ -265,12 +126,25 @@ export const HowAreYouFeelingPage = (): JSX.Element => {
           />
         </Navigation>
         <QuestionSection>
-          <ImageCheckboxes
-            id="feelings-checkboxes"
-            values={feelings}
-            label="How are you feeling?"
-            router={router}
-          />
+          <CheckboxGroup>
+            <fieldset>
+              <legend>How are you feeling?</legend>
+              <VerticalSpacing size={1} />
+              <LinkButton
+                textContent="skip this question"
+                arrow="right"
+                onClick={() => router.push('what-are-your-interests')}
+              />
+              <VerticalSpacing />
+              <ImageCheckboxes
+                id="feelings-checkboxes"
+                values={feelings}
+                contextGet={howAreFeelingGet}
+                contextToggle={howAreFeelingToggle}
+              />
+            </fieldset>
+          </CheckboxGroup>
+          <VerticalSpacing />
         </QuestionSection>
 
         <StyledLink as="button" onClick={nextPage}>
